pikl-kb-processor/requirements.txt
anthropic>=0.39.0
python-dotenv>=1.0.0
requests>=2.31.0
click>=8.1.7
rich>=13.7.0
pandas>=2.2.0
numpy>=1.26.0
sentence-transformers>=2.3.0
tqdm>=4.66.0
pydantic>=2.5.0
